Remove platform-specific code in test file
The code which only enabled the Run_Until_Close_Notify test on Windows platform has been removed. This change makes the test framework more platform-agnostic and allows this particular test to be run on any operating system. This elimination of platform specificity increases the overall versatility and utility of the tests.