Add `HttpClientUtility` and more extensions method for fluent rule configuration (#114)
* Rename method in AbsoluteUriFilter.cs

* Rename 'ReplyWithJson' methods in MockedResponseAction

Changed the naming of 'ReplyWithJson' and 'ReplyWithJsonFile' methods in the MockedResponseAction class to 'ReplyJson' and 'ReplyJsonFile' respectively. This change was made to improve readability and reduce verbosity in the method names.

* Update PathFilter and add PathFilterExtensions for fluent rule configuration

* Add HttpClientUtility and related tests

A new utility class, HttpClientUtility, is introduced to create HttpClients connected to a Fluxzy instance. This class further enables the creation of HttpMessageHandler and HttpClient instances with specified endpoints and settings.

* Update URI matching method in sample code

---------

Co-authored-by: fluxzy-ci <admin@fluxzy.io>