(Tests) Add unit tests (#140)
* Remove Fluxzy directory from codecov ignore list

The 'src/Fluxzy' directory was mistakenly added to the codecov ignore list. This commit corrects this by removing it from the list, ensuring that this directory is now considered in the code coverage reports.

* Add Fluxzy component to codecov.yml

This commit includes the Fluxzy main component in the codecov file, defining its paths for tracking. It helps to accurately measure the test coverage for the overall Fluxzy project.

* Add new test data

* Add unit tests for SearchTextFilter

New tests have been added for SearchTextFilter in the Fluxzy application. These tests assess a variety of scenarios, including searching in the request body, request header, response body, and response header.

* Add 'with-request-payload.fxzy' to Fluxzy.Tests

This update includes the addition of 'with-request-payload.fxzy' to the Fluxzy.Tests project. The file has been marked to preserve the newest copy when it is output.

* Remove LegacyPcapWriter class

The LegacyPcapWriter class was removed from the Fluxzy.Core.Pcap project. This was a deprecated class handling specific pcap writing activities, which are now handled by more efficient or updated components within the system. This aims to improve the overall efficiency and maintainability of the code.

* Add NSubstitute package to Fluxzy.Tests.csproj
.

* Refactor ResponseSuggestedExtension method in HeaderUtility class

The ResponseSuggestedExtension method in HeaderUtility.cs file was refactored to avoid redundant code and promote reusability. The response type check was replaced with a call to a new method that performs the same task, reducing code length and improving readability.

* Fix bad GetHashCode that prevents equality
.

* Add HeaderUtility unit tests

This commit introduces unit tests for the HeaderUtility class in the Fluxzy.Tests project. These tests utilize example content types and validate returns for the following methods: GetRequestSuggestedExtension, GetResponseSuggestedExtension, and GetSimplifiedContentType.

* Add equality tests for various classes

This commit introduces specific equality tests for `Authority`, `FluxzyEndPoint`, `ProxyBindPoint`, `SystemProxySetting` and `Tag` classes. These tests utilize a base testing class `EqualityTesterBase` which provides a standard test structure and enforces equality and non-equality checks for each class.

* Update codecov.yml to ignore debugging code

In the updated configuration for Codecov, debugging related files under "src/Fluxzy.Core/Clients" and "src/Fluxzy.Core/Misc/Streams" directories are added to be ignored. This means these files are no longer considered in code coverage statistics.

* Add ArchiveReader unit tests

This commit introduces unit tests for ArchiveReader in the Fluxzy.Tests project. Each test thoroughly inspects different aspects of the ArchiveReader's functionality and assures that all components are properly working. Separate tests have been also created for FluxzyArchiveReader and DirectoryArchiveReader.

* Add additional assertions in EqualityTesterBase unit tests

These changes refine the unit tests by adding more assertions for equality and inequality in the EqualityTesterBase class. This includes checks for object equivalence and verifies that a string labeled "notEqual" is also recognized as not equal.

* Add ClientErrorEquality unit test
.

* Remove BOM

* Add ImportEngineProvider unit tests

A new ImportEngineProvider unit test file has been added to improve the code coverage. Changes also include minor adjustments in the Fluxzy.Tests.csproj and ImportEngineProvider.cs files, such as copy-to-output directory options and coding style.

* Add unit tests for Agent and AuthorityInfo equality

This commit introduces two new unit tests. These tests are for the equality operators of the Agent and AuthorityInfo classes. They ensure that instances of these classes are properly evaluated for their equivalence or lack thereof.

* Increase timeouts in proxy and OS detection

The timeout in AddHocConfigurableProxy has been increased from 20 to 40 to allow for more flexibility. Timeout for OS detection under OSX platform has also been extended from 40 to 60, allowing more time for slower systems and preventing premature termination.

---------

Co-authored-by: fluxzy-ci <admin@fluxzy.io>