(Tests) Increase global coverage (#144)
* Improve Equality tests coverage

* Add ArchivingPolicyEquality

* Add more test for `EqualityTesterBase`

* Fix hostname GetHashcode

* Improve equiality test

* Add test for `Certificate`

* Add `ExportToPem` test

* Centralize test certificate passphrase

* Remove SessionIdentifier from ProxyExecutionContext

* Add more check on `ProducerContext`

* Add FluxzyArchiveReaderTests and update ArchiveReaderTests

A new FluxzyArchiveReaderTests file has been added to set up tests for the FluxzyArchiveReader. In ArchiveReaderTests, additional assertions have been added to the tests for better error checking and a new test method Validate_Metric_Builder has been introduced to validate different aspects of the exchange metric information.

* Add CurlExportFolderManagement unit tests

A new file `CurlExportFolderManagementTests.cs` was added under the `/Fluxzy.Tests/UnitTests/Util/` directory. This file contains unit tests for the CurlExportFolderManagement class, ensuring relevant functionality - such as saving to a file stream - works correctly.

* Add Traces directory to codecov ignore list

A new directory "Traces" has been added to the ignore list in the codecov configuration file. This update ensures that the code coverage tool will no longer report on anything within the Trace directory.

* Add throwOnFail parameter to QuickRun methods

In the ProcessUtils class, a new boolean parameter, "throwOnFail", has been added to the QuickRun and QuickRunAsync methods. This optional parameter allows for configuration of whether an exception should be thrown upon failure.

* Add unit tests for ProcessUtils

This commit introduces various unit tests for the ProcessUtils functionalities. The tests cover different cases such as success, failure, and output expectations for the QuickRunAsync and QuickRun methods.

* Remove unused QuickSlim class

The QuickSlim class located in Fluxzy.Core/Misc directory was deleted. This class was unused and its removal cleans up the code base, making it easier to maintain.

* Add equality tests for SearchStreamResult

Implemented a new unit test specifically designed to assess the equality of SearchStreamResult objects. The test includes conditions for both equal and unequal scenarios for thorough validation.

* Remove CollectionExtensions and redundant AsStream method

Deleted the CollectionExtensions.cs file as it was no longer needed. Also simplified the StreamUtils.cs by removing the redundant AsStream method, as it duplicated existing functionality elsewhere in the codebase.

* Add CA verification unit test to CertificateTests

A new test method, "IsCa", has been added to the CertificateTests class. The method verifies if the built-in certificate from FluxzySecurity is indeed a Certificate Authority (CA). This addition strengthens our unit testing coverage for the certificate utility functions.

* Add NetworkInterfaceInfoProvider unit tests

This commit introduces unit tests for the NetworkInterfaceInfoProvider. It ensures that the GetNetworkInterfaceInfos method returns non-null IP addresses and interface names for every network interface.