Expose write events to .NET API consumer (#153)
* Increase access level of property in Proxy.cs

The access level of the 'Writer' property in the Proxy.cs file has been changed from internal to public. This change also includes the addition of a brief summary for the property, which will be visible in intellisense and documentation.

* Add method documentation to RealtimeArchiveWriter

In the RealtimeArchiveWriter class, comments detailing what each method does were added. Each summary includes a concise explanation of the method purpose, the expected parameters, and potential return values where applicable. Included descriptions for the class itself and the ArchiveUpdateType enum as well.

* Add Short timeout constant to TimeoutConstants

The Short timeout constant has been added to the TimeoutConstants class. This new constant will return 15 if the current platform is OSX, and 10 otherwise. This will help the test cases to run with appropriate timeouts based on the OS platform.

* Add EventWriterTests for Fluxzy

A new file named EventWriterTests.cs was added in the UnitTests directory. This file contains tests for the EventWriter class of the Fluxzy library. In particular, it provides a test for checking the completion of an exchange update through a proxy.

---------

Co-authored-by: fluxzy-ci <admin@fluxzy.io>