Add buffer margin to HTTP11HeaderBlockReader
Introduced a new parameter, BufferMargin, to HTTP11HeaderBlockReader and added a checking condition to automatically increase the size of the buffer if the total read quantity along with the buffer's margin exceeds the buffer's length. This addition aims to avoid resizing the buffer when adding control headers to the client.