Fix insufficient buffer when original request is a few bytes below the threshhold (#164)
* Add buffer margin to HTTP11HeaderBlockReader

Introduced a new parameter, BufferMargin, to HTTP11HeaderBlockReader and added a checking condition to automatically increase the size of the buffer if the total read quantity along with the buffer's margin exceeds the buffer's length. This addition aims to avoid resizing the buffer when adding control headers to the client.

* Add RequestProcessingBuffer unit tests

The new tests focus on testing the buffer limit of the RequestProcessingBuffer class in the Fluxzy project. The test suite includes a number of test cases to verify that the buffer limit behaves as expected when it is around its set limit.