Allow fluxzy to forward outbound requests to another HTTP proxy (#202)
* Add MemoryUtility class for memory operations

A new class, MemoryUtility, has been added to Fluxzy.Core/Utils. This class provides utility methods for memory operations including copying formatted data and shifting destination buffers. This implementation greatly aids in more convenient data transformations and buffer manipulations.

* Add ProxyConfiguration class

A new class called ProxyConfiguration has been introduced in the Fluxzy.Clients namespace. This class is responsible for handling the configuration of a proxy server, which includes host information, port number, and the Proxy Authorization header.

* Add UpstreamProxyManager and UpstreamProxyConnectResult

This commit introduces two new classes to handle upstream proxy connections and connection results. The UpstreamProxyManager handles the connection to upstream proxy servers, while the UpstreamProxyConnectResult enum lists possible results of a connection attempt.

* Add unit tests for UpstreamProxyManager

This commit introduces unit tests for the UpstreamProxyManager class. These tests primarily cover two methods, WriteConnectHeader and ValidateConnect. They aim to ensure the robustness of HTTP proxy connection management.

* Handle ProxyConfiguration in fluxzy flow

* Add metrics on proxy connection

* Replace ProxyConfiguration with ConnectConfiguration

In the `UpstreamProxyManager` and `UpstreamProxyManagerTests` classes, the `ProxyConfiguration` parameter has been replaced with `ConnectConfiguration`. This shift was reflected in the `WriteConnectHeader` and `Connect` methods in the `UpstreamProxyManager` class, as well as in the `WriteConnectHeader` and `ValidateConnect` methods in the `UpstreamProxyManagerTests` class.

* Add proxy configuration support in connection builder

The commit introduces a ConnectConfiguration struct for proxy settings and modifies the DnsUtility and IRemoteConnectionBuilder for proxy support. It also includes timing for the proxy connection process, aiding in performance monitoring and debugging.

* Update proxy configuration assignment

Added null-check and zero-check for Host and Port before assigning a new ProxyConfiguration. This prevents incorrect configurations from being set up and subsequent potential issues in the system.

* Add YamlIgnore attribute to NoEditableSetting property

This change was done to ensure that the NoEditableSetting property in the Action class is ignored when generating YAML configurations. The attribute will prevent any unnecessary modifications that could potentially cause issues.

* Add tests for WithRuleUpStreamProxyAction class

This commit introduces new unit tests for the WithRuleUpStreamProxyAction class, thoroughly testing and validating its functionality. It includes cases for both mocked scenarios and pass-through functionality, enhancing the robustness of the project's test suite.

* Remove `noEditableSetting` from action docs

The `noEditableSetting` has been removed from various action documentation files. The commit includes changes in `SkipSslTunnelingAction.md`, `SkipRemoteCertificateValidationAction.md`, `ForceHttp11Action.md`, `RemoveCacheAction.md`, `AbortAction.md`, and `MountWelcomePageAction.md`. Additionally, the `searchable-items.json` has been updated.

* Add UpStreamProxyAction documentation

This commit introduces a new file for the UpStreamProxyAction documentation. It includes a detailed description, properties table, usage examples, .NET reference, and related action links. The documentation is essential to understand the function and configurable settings of this action.