Publish fluxzy on docker.io: https://hub.docker.com/r/fluxzy/fluxzy (#203)
* Add command line args parsing

* Update environment variable fetching in ContainerEnvironmentHelper

The "FLUXZY_MODE_REVERSE_PORT" environment variable fetching has been updated in the ContainerEnvironmentHelper.cs file. The method TryGetEnvironmentVariable is changed to TryGetInt32EnvironmentVariable to explicitly indicate the type of the variable. The value is now being added as a string to the final arguments list.

* Add unit tests for ContainerEnvironmentHelper

This commit introduces unit tests for the ContainerEnvironmentHelper within the Fluxzy.Tests.UnitTests project. These tests primarily focus on testing 'IsInContainer' and 'CreateArgsFromEnvironment' methods, validating the correct behaviour based on a variety of different inputs and scenarios.

* Add DockerFile

* Read container settings from EV instead of args

* Use alpine image instead of debian:bullseye, reduce container image to less than 100MB

* Add correct entry point to allow using shell inside the container

* Cleanup SignHelper

* Remove docker docs of default docs path

* Update `.dockerignore` to exclude build pipeline

* Add `DockerHelper`, building docker image and pushing to docker.io

* Add `fluxzy-cli-publish-docker` pipeline option to publish docker image to docker.io

* Add a docker-compose.yml example

* Add README.md for docker hub

* Cleanup last added files

* Add release-docker workflow

* Add link to docker hub on default release