Add subcommand `default` on `fluxzy certificate` to get or set the user level certificate (#212)
* Update certificate initialization in CertificateCommandBuilder

Prevent failing if default cert is invalid.

* Remove nullability from certificateProvider parameter

The nullability from the 'certificateProvider' parameter in 'GetSourceProvider' method has been removed. It should now ensure that a valid value must always be provided when calling this method.

* Make FluxzySecurity non-static to allow more testing

* Move environment-related classes to Core project

The DictionaryEnvironmentProvider, EnvironmentProvider, and SystemEnvironmentProvider classes have been moved from the Dockering subproject to the Core project. This change helps centralize and organize code that is related to managing environment variables, making it more accessible to other subprojects if needed.

* Add certificate setting and expand variable for FluxzySecurity

Refactored FluxzySecurity by adding support for the setting of certificates, the creation of necessary directories, and expansion of environment variables. Enhanced EnvironmentProvider by providing resolution for environment variables. Also introduced unit tests to validate the new functionality.

* Add unit test for `default` command