Make running tests in Linux possible with non-root user (#243)
* Replace 'Cancel' function with 'CancelAsync' in Program.cs

This commit replaces the Synchronous 'Cancel' function with the Asynchronous 'CancelAsync' function to allow task cancellation to happen asynchronously in the Fluxzy.Core.Pcap.Cli/Program.cs file. This change ensures that the program doesn't get held up waiting for tasks to be cancelled and hence improves efficiency.

* Refactor executable command name retrieval

The method for generating the executable command name has been moved from FluxzyNetOutOfProcessHost to a new helper class, FluxzyNetOutOfProcessHostHelper. This aids in keeping the main class clean and makes the same functionality accessible for other components if needed in the future.

* Update permission check in ProcessUtils

The permission check in the ProcessUtils class now considers both the 'CanElevated' function and the new 'HasCaptureCapabilities' function. This change was made to increase the flexibility and accuracy of our permission checks.

* Add capability check to ProcessUtilX class

Two new methods were added to the ProcessUtilX class in the Fluxzy.Core module. The 'CanElevated' function checks if the current process can be elevated with 'sudo'. The 'HasCaptureCapabilities' function verifies if the current process has specific capabilities. It now also includes a private readonly HashSet with required capabilities.

* Reuse property

* Add capability check unit test for Linux only

* Add Utility.cs in Fluxzy.Tests namespace

A new Utility.cs file has been included in the Fluxzy.Tests namespace. This utility class contains an async function, AcquireCapabilities, used mainly for adjusting the permissions of an executable. The function checks and sets required network capabilities for the input executable.

* Set capabilities of `fluxzynetcpa` before running test on Linux

* Remove unused HasCapabilities unit test