Fix and optimize behaviors around  `Connection` header and re-used transport connections (#444)
* Fix complete event not raised with close + fixed content length

* Better variable name

* Fix complete not raised when no expected length

* Avoid closing downstream connection systemically

* Fix dns overload test

* Remove again example.com from samples

* Handle 0-length Connection close

* Add unit test

* Replace SetResult with TrySetResult in Http11PoolProcessing

Updated the ExchangeCompletionSource to use TrySetResult instead of SetResult. This change allows for a safer attempt to set the result without throwing an exception if the result has already been set.

* Improve connection handling and resource disposal

Added a `_disposed` field to manage object state and prevent operations on disposed objects. Updated the `Dispose` method to check this state before proceeding. Enhanced connection management by replacing asynchronous client reading with a loop to handle incoming connections more effectively. Ensured errors are ignored during listener shutdown and that connections are properly completed before resource release.

* Initialize shouldClose to false in ResponseHeader.cs

Added a line to set the `shouldClose` variable to false
before returning the result of the status code check.
This change ensures that the connection closure logic
is consistently handled in this context.

* Remove connection closure check in exchange completion

The check for whether to close the downstream connection has been removed. The code now simply awaits the completion of the exchange without modifying the `shouldCloseConnectionToDownStream` variable.

* Update read completion condition in MetricsStream.cs

Modified the condition for checking if reading is complete in `MetricsStream.cs`.
The check now verifies if `TotalRead` is greater than or equal to `_expectedLength` instead of the previous logic.
This change ensures correct evaluation of the final read notification when the expected length is specified.

* Added error handling in ReadAsync to manage TLS connection closure without notifications

* Enhance connection handling and response processing

- Added checks in `Http11PoolProcessing.cs` to manage connection closure based on response body presence and `Content-Length` header.
- Introduced `ReadUntilClose` property in `Exchange.cs` to control reading until connection closure.
- Updated `ProxyOrchestrator.cs` to close downstream connections when `ReadUntilClose` is true.
- Removed outdated comment regarding HTTP/2 response handling.

* Refactor TransformResponse to use async function

Updated the TransformResponse method to utilize an asynchronous function. The new implementation now waits for the originalStream to be drained asynchronously before creating and returning the MemoryStream with the drained content.

* Update ConnectionCloseTestServer and add new tests

- Added _emptyBody and _ommitContentLength fields to ConnectionCloseTestServer.
- Updated TcpListener port from 0 to 8501 and modified constructor.
- Implemented checks for empty body and missing Content-Length in HTTP responses.
- Created RandomConnectionBehaviourTest.cs with tests for connection closure scenarios.
- Removed SubsequentSslClose.cs as it is no longer needed.

* Remove bad env var

* Set HTTP handler timeout to 15 seconds

The method now returns a new object with a configured timeout of 15 seconds for the HTTP message handler, improving request management by adding a time limit for responses.